Move tests to a separate directory.
